+++ CHANGELOG.md
@@ -12,7 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `git fetch` command
 - `git pull` command
 - Singular aliases for `groups` and `repos` commands
-- `git branches` commands: `list`
+- `git branches` commands: `list`, `checkout`
+- `--ref` flag for `status` command
+
+### Changed
+
+- `OK` status to `Clean`
+- Replace `--ok`, `--dirty` and `--missing` flags with `--show` in `status` command
 
 ### Fixed